+++ src/main/kotlin/uk/q3c/krail/core/form/FormTableSection.kt
@@ -3,8 +3,14 @@ package uk.q3c.krail.core.form
 import com.vaadin.event.selection.SelectionEvent
 import com.vaadin.event.selection.SelectionListener
 import com.vaadin.ui.Grid
+import uk.q3c.krail.i18n.CurrentLocale
+import uk.q3c.krail.i18n.Translate
 
 class FormTableSection<BEAN : Any>(val form: Form, override val rootComponent: Grid<BEAN>, val dao: FormDao<BEAN>) : FormSection, SelectionListener<BEAN> {
+    override fun translate(translate: Translate, currentLocale: CurrentLocale) {
+        rootComponent.locale = currentLocale.locale
+    }
+
     override var mode: EditMode = EditMode.READ_ONLY
 
     init {
@@ -20,6 +26,5 @@ class FormTableSection<BEAN : Any>(val form: Form, override val rootComponent: G
             }
         }
     }
-
 
 }