+++ src/main/kotlin/uk/q3c/krail/core/i18n/CommonLabelKey.kt
@@ -6,15 +6,32 @@ import uk.q3c.krail.i18n.I18NKey
  * Created by David Sowerby on 26 Aug 2018
  */
 enum class CommonLabelKey : I18NKey {
+    About,
     After,
     Before,
+    Cancel,
+    Edit,
+    Error,
     Full_Screen,
+    Help,
     Home,
+    Large,
+    Loading_,
     Menu,
-    Log_Out,
     Notifications,
     No,
+    Ok,
+    Push,
+    Save,
+    Saved,
     Settings,
+    Small,
     Title,
+    Warning,
     Yes
+}
+
+enum class UserStatusKey : I18NKey {
+    Log_In,
+    Log_Out
 }