+++ src/main/kotlin/uk/q3c/krail/core/view/component/UserStatusPanel.kt
@@ -15,8 +15,34 @@ package uk.q3c.krail.core.view.component
 import com.vaadin.ui.Button
 import com.vaadin.ui.Component
 import com.vaadin.ui.Label
+import java.io.Serializable
 
+
+/**
+ * Provides a component representing the user's current login status, and a button to log in / log out as appropriate
+ * Responds to changes caused by the user logging and out
+ *
+ *
+ */
+@Deprecated("replaced by UserStatusComponents")
 interface UserStatusPanel : Component {
+    val usernameLabel: Label
+    val login_logout_Button: Button
+
+    val actionLabel: String
+    val userId: String
+}
+
+/**
+ *
+ * Provides a component representing the user's current login status, and a button to log in / log out as appropriate
+ * Responds to changes caused by the user logging and out
+ *
+ * Replaces, and is functionally the same as [UserStatusPanel], except that this is not a component.  Rather than create
+ * a layout for the two components (which may be wasted effort), this approach assumes that the constituent components
+ * will be put wherever the developer wants them.  The implementation maintains the relationship between those components
+ */
+interface UserStatusComponents : Serializable {
     val usernameLabel: Label
     val login_logout_Button: Button